+++ Frontend/src/components/AddRentalObjects/AddImageForm.vue
