+++ Frontend/src/components/AddRentalObjects/AddRentalObjectForm.vue
@@ -27,7 +27,7 @@
 
 <script>
 
-import Calendar from '../components/Calendar.vue'
+import Calendar from '../Utils/Calendar.vue'
 
 
 export default {
