+++ Frontend/src/components/Booking/Receipt.vue
