+++ Frontend/src/components/Booking/ReceiptList.vue
