+++ Frontend/src/components/RentalDetails/CitiesList.vue
