+++ Frontend/src/components/RentalDetails/City.vue
