+++ Frontend/src/components/RentalDetails/DisplayHotAmenity.vue
