+++ Frontend/src/components/RentalDetails/RentalImages.vue
