+++ Frontend/src/components/RentalDetails/RentalObject.vue
