+++ Frontend/src/components/Searchbar/SearchBar.vue
@@ -35,7 +35,7 @@
 </template>
 
 <script>
-import Calendar from './Calendar.vue'
+import Calendar from '../Utils/Calendar.vue'
 
 export default {
   components: {
