+++ Frontend/src/components/Searchbar/SearchBarModal.vue
@@ -53,7 +53,7 @@
 </template>
 
 <script>
-import Calendar from './Calendar.vue'
+import Calendar from '../Utils/Calendar.vue'
 
 export default {
   components: {
