+++ Frontend/src/components/User/AddUserForm.vue
@@ -44,7 +44,7 @@
 </template>
 
 <script>
-import store from '../store.js'
+import store from '../../store.js'
 export default {
   computed: {
     validatePassword() {
