+++ Frontend/src/components/Utils/AmenityLogo.vue
