+++ Frontend/src/views/Home.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import CitiesList from '../components/CitiesList.vue'
+import CitiesList from '../components/RentalDetails/CitiesList.vue'
 
 export default {
   components: {
