+++ Frontend/src/views/Overview.vue
@@ -5,7 +5,7 @@
 </template>
 
 <script>
-import Filter from '../components/Filter.vue'
+import Filter from '../components/Utils/Filter.vue'
 export default {
   components: {
     Filter,
